refactor: 클래스 이동
Main 클래스와 Input, Output 클래스를 MVC 디자인 패턴에 맞게 이동
그에따른 Import 문 수정
