+++ src/test/java/OutputTest.java
@@ -2,7 +2,7 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import racinggame.controller.Output;
+import racinggame.domain.Output;
 
 public class OutputTest {
     Output output;
