+++ src/test/java/RacingGameTest.java
@@ -1,13 +1,12 @@
 import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.RepeatedTest;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import racinggame.controller.Output;
+import racinggame.domain.Output;
 import racinggame.domain.Car;
-import racinggame.controller.Input;
+import racinggame.domain.Input;
 import racinggame.domain.RacingGame;
 
 import java.util.ArrayList;
